+++ docs/src/App.tsx
@@ -1,8 +1,10 @@
 import { HashRouter as Router, Routes, Route } from 'react-router-dom'
 import Header from './components/Header'
 import Home from './pages/Home'
-import OurWork from './pages/OurWork'
-import MeetUs from './pages/MeetUs'
+import WhatWeDo from './pages/WhatWeDo'
+import Work from './pages/Work'
+import HowWeWork from './pages/HowWeWork'
+import About from './pages/About'
 import Contact from './pages/Contact'
 
 function App() {
@@ -12,8 +14,10 @@ function App() {
         <Header />
         <Routes>
           <Route path="/" element={<Home />} />
-          <Route path="/our-work" element={<OurWork />} />
-          <Route path="/meet-us" element={<MeetUs />} />
+          <Route path="/what-we-do" element={<WhatWeDo />} />
+          <Route path="/work" element={<Work />} />
+          <Route path="/how-we-work" element={<HowWeWork />} />
+          <Route path="/about" element={<About />} />
           <Route path="/contact" element={<Contact />} />
         </Routes>
       </div>